Add `dmft-ipt` submodule
The current tip of the lattices branch in bellomia/MOTTlab has proven to be capable of treating the MIT on the square lattice, at the DMFT/IPT level. Hence we include it here as our ipt submodule. When the lattices branch will be merged into MOTTlab master we should move this submodule accordingly I suppose, but for now let's stay in the lattices branch.