SSRs do spoil (in some sense) additivity!
What am I committing here?

> just a try at modifying the superslection_factor figure, to include
  a panel for S2, with the naive idea to make evident the cancellation
  between a huge factor for S1 and huge factor for S2, giving the nice
  almost identity factor for 2S2-S1. This just not works and there are
  profound reasons behind...hence this update would never be merged to
  master and will stay here as a memento.

What I mean?

> if I apply a SSR to the total bond correlation, which is a mutual info
  I could naively expect that, given MI = 2S1 - S2, the SSR would give:

    MI|SSR = 2 * S1|SSR - S2|SSR

  This notation is truly misleading and could lead to severe paradoxes,
  like thinking that SSR and algebraic operations do not commute.

How is this solved?

> changing how we phrase and write the terms, i.e. we can write

  MI|SSR = 2 * S1|SSR - S2|SSR

  but we need to make sure to understand that S1|SSR is NOT the local
  entanglement and S2|SSR the quasilocal entanglement under the SSR.

————————————————————————————————————————————————————————————————————————

LET'S CLARIFY THE CONCEPT...

Elaborating a bit I believe we may want to keep a clear cut between the
mathematical and the physical names.

- Von Neumann entropies and mutual information are mathematical objects,
  and the latter is defined in terms for the former.

- Local (and quasilocal) entanglement and total correlation are physical
  quantities, that happen to correspond to one mathematical object under
  some assumptions, but to some other thing under other assumptions.

Without SSRs everything is simple cause the (quasi)local entanglement is
given by the relevant von Neumann entropies of the relevant RDMs and the
total correlation is the mutual information between two orbitals, hence
we have additivity in the sense that the total correlation comes from an
algebraic sum of local and quasilocal entanglement.

With SSRs there's care to be taken since local & quasilocal entanglement
are not von Neumann entropies, yet the total correlation is still defined
as a mutual information, acting on a filtered RDM.
Hence it comes from an algebraic sum of von Neumann entropies, but not of
entanglement measures, since they are not von Neumann entropies.

—————————————————————————————————————————————————————————————————————————

LET'S TRY TO FIND AN IMPROVED NOTATION THEN...

So, let's call E1 the measure of entanglement between a single site and
the whole. S1 is instead just the mathematical definition of von Neumann
entropy of the RDM1 (same for E2, S2 and RDM2 for the dimer quantities,
instead of single site).

Whithout SSRs we have that E1 = S1, E2 = S2 and MI = 2S1 - S2, hence the
algebra works fine cause you literally have MI == 2E1 - E2.

But with SSRs it's not true that E1 = S1 and E2 = S2.
For E2 I'm not sure, we would need to derive something following the
ideas in that JCTC for the dimer case.
But for E1 the formula is definitely not S1|{SSR on the RDM1}.
It is a completely different formula than the von Neumann entropy acting
on the untouched RDM for the single site.
It can be derived of course as coming from imposing the SSR on the whole
DM, but this is irrelavant here.

That's why when you compute MI with the SSR you do something to the RDM2
and then get some S1|SSR and S2|SSR to plug in the MI|SSR formula, but
you don't have additivity with E1|SSR and E2|SSR, cause these two aren't
von Neumann entropies anymore.
Why the dimer and single site entanglement are not von Neumann entropies
anymore? Because applying SSRs to the whole DM you make it in general an
impure state and von Neumann entropies of its reduced systems are not
measures of entanglement anymore.

————————————————————————————————————————————————————————————————————————

FINALLY A COMMENT ON WHY S1|SSR IS EXACTLY THE SAME AS S1

This is another source of confusion... S2|SSR is different from S2, but
not a correct expression for E2|SSR, ok. Further, while we know there is
an inequality associated to the SSRs (E2|SSR ≤ E2) this is not valid for
S2 and S2|SSR.

But having S1|SSR == S1 can become even more confusing... we are doing a
restriction on the RDM2 and this leads to a different S2, but how can it
leave unchanged S1, a functional of RDM1, which is a trace of RDM2?

Here my best (technical) answer:

      We know SSRs are local. Hence each local N or parity has to be
      conserved. When we trace from either an untouched or SSR RDM2,
      we need to keep the "to be traced site" state fixed. Hence we
      cannot mix up its local N or parity. Thus the elements we set
      to zero by applying the SSRs never enter the trace defining the
      single site RDM.