Revert "Update BOND_ENTANGLEMENT figure"
This reverts commit ebf36a1263454d8481507d84acbee98ad049c032.

We finally understood that no, the upper bound cannot be halved. In the
limit in which the composed system AB is pure, indeed we have

I(A:B) = 2*E(A:B) = 2S(A) = 2S(B)

And the factor of two is NOT a matter of convention / unit of measure.

In fact, it can be proven that for a pure AB system the so-called
/quantum discord/ is exactly equal to the entanglement, so that

I(A:B) = E(A:B) + D(A:B), where D is the discord.

But in the general case one still has

I(A:B) ≥ E(A:B) + D(A:B) + C(A:B) (where C is the classical correlation)

and there is no theorem regarding how E and D are related. So that the
equality condition is not found in the trivial case in which C = 0 and
one has I = 2E = E + D, but in some nontrivial case in which C = D = 0.
We cannot assume D = E, outside the pure AB case, neither we can assume
that D = 0 for our system, so we are stuck with I(A:B) ≥ E(A:B).